(+) Specify how posts are encrypted
This is done using a „broadcast key“ so that posts have to be encrypted
only once for all friends. This also allows caching.
(/=) fix in the Discovery protocol : avoid a reflection attack that
would make A believe that B has exactly the same friends as A.
(/=) Include the trusted T in the Exchange messages 4, 5 to avoid
exploits of the fact that any peer can act as the server. (in contrast
to the original NS-Loewe protocol setting)
(+) Alternative Exchange protocol draft (not sure if useful)